chore(clients): move client files to their own module dir
