+++ src/clients/bitcoind.js
@@ -2,7 +2,7 @@ import axios from "axios";
 import BigNumber from "bignumber.js";
 import { bitcoinsToSatoshis } from "unchained-bitcoin";
 
-async function callBitcoind(url, auth, method, params = []) {
+export async function callBitcoind(url, auth, method, params = []) {
   // FIXME
   // eslint-disable-next-line no-async-promise-executor
   return new Promise(async (resolve, reject) => {
